+++ CHANGELOG.md
@@ -9,6 +9,18 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Add `Value.Bytes()`
+
+### Changed
+
+- **[BC]** Rename `Value.IsEmpty()` to `IsZero()` to better distinguish between unpopulated values and empty content
+- **[BC]** `Value.AsString()` and `AsBytes()` now return an error if the `Value` is the zero-value
+- **[BC]** `Value.String()` now panics if the `Value` is the zero-value
+
 ## [0.1.2] - 2019-11-20
 
 ### Added